+++ Composer/packages/extensions/adaptive-form/src/components/fields/QnAActionsField.tsx
@@ -9,7 +9,7 @@ import { Link } from '../Link';
 
 export const QnAActionsField: React.FC<FieldProps<string>> = function StringField() {
   const { currentDialog, projectId } = useShellApi();
-  const qnaUrl = `/bot/${projectId}/qna/${currentDialog.id}`;
+  const qnaUrl = `/bot/${projectId}/knowledge-base/${currentDialog.id}`;
   const content = formatMessage('Go to QnA all-up view page.');
   return <Link href={qnaUrl}>{content}</Link>;
 };
